# apps/firescrapling/backend/openapi.yaml
openapi: 3.1.0
info:
  title: FireScrapling API
  version: 1.0.0
paths:
...
  /v1/billing/checkout:
    post:
      tags:
      - billing
      summary: Billing Checkout
      operationId: billing_checkout_v1_billing_checkout_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Billing Checkout V1 Billing Checkout Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CheckoutRequest:
      properties:
        plan:
          type: string
          title: Plan
          description: pro or team
        success_url:
          type: string
          maxLength: 2048
          minLength: 8
          title: Success Url
        cancel_url:
          type: string
          maxLength: 2048
          minLength: 8
          title: Cancel Url
      type: object
      required:
      - plan
      - success_url
      - cancel_url
      title: CheckoutRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError